app/requirements.txt
Flask==3.0.0
Flask-Login==0.6.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Limiter==3.5.0